Made a semi functional API using the Gino library
The idea was to preserve the SQLAlchemy definition of ODM2 while switching to
asyncpg as the database driver. gino seemed to be the best way to do this.

However I think if we want the performance of asyncpg we have to get rid of
SQLAlchemy/Gino I think they are of dubious benefit anyways. I will commit this
code as is in case I want to use it anyways. Future commits will probably rip
out Gino/SQLAlchemy.