Add samplingfeature post method
Asyncpg is configured to map from shapely objects (in python) to postgis
objects. Since the api can be scaled horizontally and the database cannot it is
good practice to do conversion from wkt to a binary gis object in the python
side of things. This means the database will do the minimal amount of
operations we can get away with.

It should be possible to switch out shapely for something more high level like
geopandas fairly easily. I made a decision together with @jamessample to
default all gis colums currently in use to the 4326 reference frame which is
basically lat-long.