Added projects to ODM2 api
After talking with @JamesSample we settled on using the 'directives'
table. The alternative would have been to use the 'dataset' table to
keep track of projects. This would have been a little more direct but
also a little more hackish. We think this is be a good way to do it.

Also added --no-ignore-rbac flag to remove modifications endpoints in
graphiql.