Added annotations endpoint
The annotations endpoint is needed as described in #25 to hold metadata
for results and methods. Since we want to store json data as annotations
an extra column had to be added to the annotations table.

Adding this column using an 'ALTER TABLE' statement led @zofiarudjord
and me to discover 2 bugs in db_initiate.py which are now fixed.

We also added an optional List property to methods that can be used to
add annotations directly on a method.