Changed logic for getting stored people/affiliations #43
Fetching stored person based on user email instead of port-id. This
solves issues when people row was populated from on-prem active
directory export (see #38)

Also removed the concept of external identifier to niva-port. This makes
odm2 unaware of whether the person was populated from a logged in user
or other sources. This may be a drawback of this approach, but I think
it also makes the system more robust by just referring to users with
their UserPrincipalName (email)

changes:

- removed niva-port as externalidentifiersystem in metadata script
- fetch user based on affiliations.primaryemail instead of
externalidentifiersystem