Expanded the FeatureActionsCreate schema to be more general
A feature_action row requires an actionid and a samplingfeatureid. Since
it proved cumbersome to keep track of samplingfeatureid's the schema
allowed for uuids instead and finds the associated samplingfeature.

Since both `samplingfeatureuuid` and `samplingfeaturecode` have a unique
constraint it makes sense to allow either when trying to find a
samplingfeatureid that should be referenced in a feature_action row. If
both a uuid and a code are supplied they have to both match to be
valid. If only one is supplied that one is simply used to query for a
samplingfeature.