Writing new observations to aquamonitor API
- Created API wrapper for aquamonitor API (aquamonitor client) based on
httpx AsyncClient

- manually constructed the API request/response types from aquamonitor
API in aquamonitor_api_types.py. Kept the same names from the API with
the exception of some fields that use leading underscore which did not
work very well with pydantic. Mapped these using aliases, but the
serialized json remains the same
- hardcoded a map between method name and nivabase id. There is no
endpoint to fetch method based on method name currently, so doing this
for now. May consider to talk with Roar to add a new endpoint in the API
for this
- Constructed a new data structure for begroing observations where we
have a list of (taxon, method, value). This is easier to work with and I
think this should be moved over to the frontend
- Roar has created a system user to interact with aquamonitor API. I
stored this as k8s secret aquamonitor-api-user

TODOS:
- I am calling a DELETE after each write. This is not 100% safe, but
also very pragmatic. May consider to opt for a safer regime when testing
- I made some helper utils for logging outgoing requests, adding
span-id/trace-id etc that should go to nivacloud-logging
- station codes are not 1:1 with nivabase station codes. I think we may
need to query the BEGROING.IMPORT_STATIONS table to get the actual
station code? this needs to be discussed
- need to actually integrate this into the endpoint. I am just calling
this manually now
- not sure what to do in such situations as we'd have two
PROJECTS_STATIONS rows in nivabase. Don't want to duplicate values so I
just throw an error for now

Related issue:
https://github.com/NIVANorge/begroing-data-upload/issues/60

- handling errors coming from aquamonitor API
  - if response is json, using message if it's there

- saving to aquamonitor as part of POST /begroing/begroing_result
- introduced a new environment variable WRITE_TO_AQUAMONITOR. IF set to
true, the following env variables are also required:

- AQUAMONITOR_USER
- AQUAMONITOR_PASSWORD

Valid values for WRITE_TO_AQUAMONITOR:

True values are y, yes, t, true, on and 1; false values are n, no, f,
false, off and 0. Raises ValueError if val is anything else.

TODO: I am mapping the data structure so that we get a list of
observations with taxon, value, method. I think we should work with this
data structure all the way up to the frontend as well. Suggest we do
this as a separate change as it involves frontend changes as well

- Added unit test for aquamonitor client

Struggled a bit with finding a good test scope for this, but landed on
mocking the http layer of the client so that I test that it does the
expected http calls. May consider to abstract away a lot of the mock
setup so that it is easier to setup for other tests.

Added aquamonitor integration test, calling the actual API

intended usage is to run in an interval in CI with credentials set. This
way we can discover regressions in the API when they happen
- added circleci config for nightly test
  - using credentials added to circleci env variables (within
odm2-postgres-api)

TODO: the credentials I am using have write privileges. Should figure
out if there is a concept for read-only users in aquamonitor API.

ref issue https://github.com/NIVANorge/AquaMonitor/issues/22

- using handle_aquamonitor_error for all responses
  - this function checks status code for errors and parses json body for
error message if it exists