Writing new observations to aquamonitor API
- Created API wrapper for aquamonitor API (aquamonitor client) based on
httpx AsyncClient

- manually constructed the API request/response types from aquamonitor
API in aquamonitor_api_types.py. Kept the same names from the API with
the exception of some fields that use leading underscore which did not
work very well with pydantic. Mapped these using aliases, but the
serialized json remains the same
- hardcoded a map between method name and nivabase id. There is no
endpoint to fetch method based on method name currently, so doing this
for now. May consider to talk with Roar to add a new endpoint in the API
for this
- Constructed a new data structure for begroing observations where we
have a list of (taxon, method, value). This is easier to work with and I
think this should be moved over to the frontend
- Roar has created a system user to interact with aquamonitor API. I
stored this as k8s secret aquamonitor-api-user

TODOS:
- I am calling a DELETE after each write. This is not 100% safe, but
also very pragmatic. May consider to opt for a safer regime when testing
- I made some helper utils for logging outgoing requests, adding
span-id/trace-id etc that should go to nivacloud-logging
- station codes are not 1:1 with nivabase station codes. I think we may
need to query the BEGROING.IMPORT_STATIONS table to get the actual
station code? this needs to be discussed
- need to actually integrate this into the endpoint. I am just calling
this manually now
- not sure what to do in such situations as we'd have two
PROJECTS_STATIONS rows in nivabase. Don't want to duplicate values so I
just throw an error for now

Related issue:
https://github.com/NIVANorge/begroing-data-upload/issues/60