made post /begroing/indicies endpoint not rely on internal hardcoded id's
- this was a problem when running API/integration tests when we wipe
database, but don't wipe primary key sequences. Then we end up with
different id's.

solving this by querying for units, processinglevels etc as part of the
endpoint. Not very elegant, but I think this is an issue with the
normalized structure of ODM2. I suggest we discuss this and figure out a
direction