Added two new endpoints to find units and variables through api
In order for both of these endpoints to be GET @zofiarudjord and I
decided not to use the UnitsCreate schema in the endpoint. The variable
endpoint naturally uses just a string in order to find the row.

We also did a small refactor to get rid of `get_unit` so that only
`find_unit` is left. This makes `find_unit` and `find_row` a little more
similar which seemed to be a little more readable.