saving to aquamonitor as part of POST /begroing/begroing_result
- introduced a new environment variable WRITE_TO_AQUAMONITOR. IF set to
true, the following env variables are also required:

- AQUAMONITOR_USER
- AQUAMONITOR_PASSWORD

Valid values for WRITE_TO_AQUAMONITOR:

True values are y, yes, t, true, on and 1; false values are n, no, f,
false, off and 0. Raises ValueError if val is anything else.

TODO: I am mapping the data structure so that we get a list of
observations with taxon, value, method. I think we should work with this
data structure all the way up to the frontend as well. Suggest we do
this as a separate change as it involves frontend changes as well