Moving mass_spec complexity from mass_spec code to api endpoints
Many of the functions to support the mass_spec workflow required hitting
several endpoints in the api. This made the code a little hard to read
and since many of these requests should actually be in a transaction it
makes sense to move this code into the api. In this way the requests
also indicate a very clear boundary of what fields are constant and
which vary.

The 4 new endpoints basically support the 4 major actions that can be
done in a mass_spec workflow.  These 4 are:
- /register_site
- /register_sample
- /register_replicas
- /register_output

    register_site
This is essentially a thin wrapper around sampling_feature_create that
either creates a new one or returns the already existing site. We could
consider not having this endpoint at all but it's nice to have them
under the same router.

    register_sample
When a physical is collected at a sewage plant (or similar) the sample
should be registered as a new samplingfeature and the 'collect' action
should be store as well.

    register_replicas
When a sample is collected it is fractionated into multiple samples that
are analyzed in the mass spectrometer. This allows for a statistical
analyses of the result. In combination with running 'blanks' through the
mass spectrometer it allows researchers to have high confidence in
detections. These fractions or replicas are registered as new 'child'
samplingfeatures by this endpoint.

    register_output
There are three types of mass_spec outputs that can be registered in the
ODM2 results. The three result types are:
- feature detection
- feature deconvolution
- feature identification
All of these output types are essentially large csv files and in the
future their contents could conceivably be stored in ODM2. For now they
will remain as files in google cloud and the result will be annotated
with their location in a google cloud bucket.