Added unit test for aquamonitor client
Struggled a bit with finding a good test scope for this, but landed on
mocking the http layer of the client so that I test that it does the
expected http calls. May consider to abstract away a lot of the mock
setup so that it is easier to setup for other tests.